Update for GetRealGOEntry function

+++ src/server/game/Server/WorldSession.h
@@ -413,7 +413,7 @@ class WorldSession
         bool IsARecruiter() { return isRecruiter; }
         uint64 GetRealGUID(uint8 packetGuid, uint8 byte, std::string ErrorMessage);
         uint64 GetRealGOGUID() { return realgoguid; }
-        uint64 GetRealGOEntry() { return realgoentry; }
+        uint32 GetRealGOEntry() { return realgoentry; }
 
     public:                                                 // opcodes handlers
